+++ sql_app/database.py
@@ -15,3 +15,9 @@ async_session = sessionmaker(
 )
 
 Base = declarative_base()
+
+
+async def get_session():
+    '''Dependency session.'''
+    async with async_session() as session:
+        yield session
